+++ src/plugins/miscellaneous/Core/src/corecliutils.h
@@ -120,7 +120,7 @@ quint64 CORE_EXPORT freeMemory();
 
 QString CORE_EXPORT digitGroupNumber(const QString &pNumber);
 
-QString CORE_EXPORT sizeAsString(double pSize, int pPrecision = 1);
+QString CORE_EXPORT sizeAsString(quint64 pSize, int pPrecision = 1);
 
 QString CORE_EXPORT sha1(const QByteArray &pByteArray);
 QString CORE_EXPORT sha1(const QString &pString);
